# docs/api-spec.yml
openapi: 3.1.0
info:
  title: Swagger Petstore - OpenAPI 3.1
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.1 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)

  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: https://petstore3.swagger.io/api/v3
...
paths:
...
  /menus/{menuId}:
    get:
      tags:
        - menus
      description: ...
    put:
      tags:
        - menus
      description: ...
    delete:
      tags:
        - menus
      description: ...
    parameters:
      - $ref: "#/components/parameters/MenuId"
...
  /customers/{customerId}:
    get:
      tags:
        - customers
      description: Get an existing customer by its id.
    put:
      tags:
        - customers
      description: Update (in full) the details of an existing customer by its id.
    delete:
      tags:
        - customers
      description: Delete an existing customer by its id.
    parameters:
      - $ref: "#/components/parameters/CustomerId"
...
components:
  parameters:
    MenuId:
      name: menuId
      in: path
      schema:
        $ref: "#/components/schemas/UUID"
...
    CustomerId:
      name: customerId
      in: path
      schema:
        $ref: "#/components/schemas/UUID"
...
  schemas:
    UUID:
      type: string
      format: uuid
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header